requirements.txt
streamlit
pypdf
Pillow
PyMuPDF
reportlab
